refactor(isTaxID): Move helper validation algorithms * General-purpose validation algorithms have benn moved to `src/lib/util/algoritms.js` to support further use. * Validation algorithms have been refactored to use strings as input * isTaxID has been refactored to use `algorithms.js`
